Plot TOF sesans data with wavelength normalisation
The SESANS polarizsation goes as P(Z) = e^(-C(Z) λ²) for wavelength λ and some
C(Z) proporational to our actual signal.  When the wavelength and C(Z)
is close to zero, then the polarisation stands in as a good proxy for
the actual sample.  On time of flight, however, it's necessary to
perform a (log P(Z))/λ² to get a real understanding of the sample.

This patch checks to see whether a data set is time of flight (by check
to see if the wavelengths are all equal) and then changes the default
plotting method to use the log over lambda square plots for time of
flight data and raw polarisation for reactor data.