Simplify model parser
The updated model parser uses only a single switch statement and most
parts contain only a single variable.  Comparatively, the old version
was execeeding the PEP 8 recommendations for both conditional depth and
number of local variables.

The trade off is that the new parser is recursive.  The maximum depth
is four, so we're not in any danger of blowing the stack, but
recussion can add its own kind of complexity.