Fix docs and code for be_polyelectrolyte.py
Docs were fixed as explained in ticket #1155 as was the code.  A
validation note was added to docs to indicate change and the fact that
full validation has not really been done.  Also all unit tests that had
non-zero salt concentration values have been commented out till we
decide how to handle the change in terms of validating unit tests.