Finish updating the Porod Model docs
clean up and adding the caveats.  It would be nice to be able to ref the
invariant and linearized fit documentation but given they are in
different repos and sasmodels in principle can be distributed without
SasView there is no sensible way to do this. Perhaps we could discuss
adding sasview docs as a reference?