Edit BCC docs and fix unit tests
Updated the BCC docs with corrections but more importantly adding the full description of the equations used as given by jonathan Gaudet in his latex writeup.  Also fixed the unit test, but it would be helpful to have an independent assesment of the correct answers. At which point some better unit test coverage could be provided. Note that this change drops the intensities by about half!